% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acondpow.R
\name{acondpow}
\alias{acondpow}
\title{Compute Approximate Conditional Power}
\usage{
acondpow(za, power, zs, infr)
}
\arguments{
\item{za}{standard normal critical value at final analysis}

\item{power}{the overall planned power of the sequential test}

\item{zs}{value of the test statistic on the standard normal scale at the
current interim analysis}

\item{infr}{information fraction (proportion of total information) at the
current interim analysis}
}
\value{
The approximate conditional power (a scalar).
}
\description{
Computes an approximation to the conditional power of a sequential test
based on the asymptotic normality of the statistic.
}
\details{
Computes an approximation to the conditional probability of the test being
rejected at the final analysis (i.e., that the test statistic will be >
\code{za}), conditional on the current data, if the alternative for which
the sequential test has power \code{power} is true.  The approximation is
based on asymptotic joint normality of the test statistic at the current and
final analysis, and on the increment in the information being independent of
the current data.
}
\examples{
acondpow(2.0, 0.85, 0.32, 0.41)

}
\seealso{
\code{\link{condpow}}; \code{\link{condpowcure}}
}
\keyword{design}
