% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin1samp.R
\name{bin1samp}
\alias{bin1samp}
\title{Sample Size for One-sample Exact Binomial Tests}
\usage{
bin1samp(p0, pa, alpha = 0.1, beta = 0.1, n.min = 20L)
}
\arguments{
\item{p0}{Null hypothesis response probability}

\item{pa}{Alternative hypothesis response probability}

\item{alpha}{Type I error rate}

\item{beta}{Type II error rate}

\item{n.min}{Minimum sample size considered}
}
\value{
\code{bin1samp} returns a vector giving the minimum sample size
(\code{n}), the critical value \code{r} (reject if outcome is more extreme
than \code{r}), the null and alternative response probabilities (\code{p0}
and \code{pa}), and the type I and type II errors (\code{size} and
\code{type2}).)

No value is returned by \code{print.bin1samp}, but it prints the input
\code{p0} and \code{pa}, the minimum sample size, the critical value
\code{r} for the test that rejects if the number of responses is \code{> r}
if \code{pa>p0} or that rejects if the number of responses is \code{< r} if
\code{pa<p0}, and the actual type I and type II error rates.
}
\description{
Determines the minimum sample size for a one-sided, one-sample exact
binomial test with specified alpha (type I) and beta (type II) errors
}
\details{
Loops over sample sizes starting with \code{n.min}, determines the critical
value for the exact test of size alpha, and calculates the type II error.
If the type II error is larger than \code{beta}, increments the sample size
by 1 and tries again.

\code{\link{print.bin1samp}} is a print method for the output of
\code{bin1samp}.
}
\examples{
bin1samp(0.9, 0.95, n.min = 100)
bin1samp(0.1, 0.05, n.min = 100)

}
\seealso{
\code{\link{print.bin1samp}}; \code{\link{pickwin}}; \code{\link{rp21}};
\code{\link{twostg}}; \code{\link{simon}}
}
\keyword{design}
