% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2.R
\name{binci}
\alias{binci}
\title{Exact Binomial Confidence Limits}
\usage{
binci(r, n, conf = 0.95)
}
\arguments{
\item{r}{number of successes}

\item{n}{total number of trials}

\item{conf}{confidence level}
}
\value{
Returns a vector containing the two confidence limits. Returns 0 for
the lower limit if r=0, and returns 1 for the upper limit if r=n.
}
\description{
Calculates exact confidence bounds on a single binomial probability
}
\details{
Calculates exact \code{1-(1-conf)/2} upper and lower confidence limits for a
binomial proportion.  The precision of the solution is the default in
\code{uniroot}.
}
\examples{
binci(54, 88)

}
\seealso{
\code{\link{twocon}}
}
\keyword{htest}
