% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binsim.R
\name{binsim}
\alias{binsim}
\title{Simulation for Power of a Sequential Fisher's Exact Test}
\usage{
binsim(
  nsamp,
  n,
  p1,
  p2,
  inf.times,
  upper,
  lower = rep(1.5, length(inf.times)),
  p.con = 0.5,
  alt = "less"
)
}
\arguments{
\item{nsamp}{Number of simulated trials}

\item{n}{Total sample size for each trial}

\item{p1}{Success probability for group 1 (control)}

\item{p2}{Success probability for group 2}

\item{inf.times}{Information times of analyses (as fractions)}

\item{upper}{Upper boundary (p-value scale, so test rejects for values
smaller than the 'upper' boundary)}

\item{lower}{Lower boundary (p-value scale)}

\item{p.con}{Proportion randomized to control (group 1)}

\item{alt}{Direction of the alternative (one of greater, less or two.sided)}
}
\value{
A vector of length \code{nsamp}, giving the test result for each
simulated trial, coded 1 if the upper boundary was crossed, -1 if the lower
boundary was crossed, and coded 0 if neither was (only the first boundary
crossing is recorded).
}
\description{
Performs a simulation to estimate the power of Fisher's Exact under group
sequential monitoring.
}
\details{
Calls \code{\link{fisher.test}}. If the experimental group (2) has a higher
event rate than the control group (1), then 'less' is the appropriate
direction for a one-sided test.
}
\examples{
binsim(20, 50, 0.1, 0.5, c(0.5, 1), c(0.05, 0.05), c(0.5, 0.5), alt = 'less')

}
\seealso{
\code{\link{b2p}}; \code{\link{fisher.test}}
}
\author{
R Gray
}
\keyword{design}
