% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correl.power.R
\name{correl.power}
\alias{correl.power}
\alias{correl.ratio}
\title{Compute Power for Failure Time Comparisons in Correlative Studies}
\usage{
correl.power(
  n,
  nd,
  r1o2,
  p1,
  acc.per,
  add.fu,
  alpha2 = 0.05,
  int = c(0.001, 5)
)
}
\arguments{
\item{n}{Total number of cases}

\item{nd}{Total number of failures}

\item{r1o2}{Ratio of the hazard in Group 1 over that in Group 2}

\item{p1}{Proportion of sample in Group 1}

\item{acc.per}{Duration of accrual}

\item{add.fu}{Additional follow-up after accrual was completed}

\item{alpha2}{Type I error rate (two-sided)}

\item{int}{Initial search interval for calculating hazard rates}

\item{power}{The desired power for the comparison}

\item{maxrat}{The maximum hazard ratio to consider in the search when
\code{greater=TRUE}. \code{1/maxrat} is the minimum ratio to consider when
\code{less=TRUE}}

\item{greater}{logical; if \code{TRUE}, find the ratio greater than 1 with
the specified \code{power}}

\item{less}{logical; if \code{TRUE}, find the ratio less than 1 with the
specified \code{power}}
}
\value{
\code{correl.power} returns a vector giving the hazard rates in
groups 1 and 2 (\code{lambda1} and \code{lambda2}), the failure
probabilities in the two groups (\code{fp1} and \code{fp2}), the standard
error of the log of the ratio of the estimated exponential failure rates
(\code{se}), the total expected number of failures under the specified
failure rates (\code{n.fail}), the normal deviate corresponding to the type
II error (\code{zbeta}), and the power (\code{power}).

\code{correl.ratio} returns a vector giving the total number of failures
(\code{n.fail}), the proportion of the sample expected to be in group 1
(\code{pr.g1}), and the hazard ratios (group 1/group 2) for which the test
will have the specified power in the directions with worse and better
outcome on group 1 (\code{r1o2.1worse} and \code{r1o2.1better}).
}
\description{
Computes power or difference for the logrank test for a correlative study
comparing failure time distributions between groups defined by two levels of
a correlative marker when there is a known amount of information (number of
events)
}
\details{
Assumes exponential failures and a clinical trial setting with uniform
accrual over \code{acc.per} units of time plus an additional \code{add.fu}
time units of follow-up, so censoring is uniform on (\code{add.fu},
\code{add.fu + acc.per}).  Test statistic is the exponential Wald test using
asymptotic normality, which is asymptotically equivalent to the logrank
test.

Only the variance under the alternative is used, consistent with
Rubinstein, Gail and Santner (1981) and \code{seqopr} and \code{srvpwr}.
\code{r1o2} can be either larger or smaller than 1.  The total failure
probability is fixed at \code{nd/n}, and hazard rates in the two groups are
computed to give the specified ratio subject to the total failure
probability constraint.  The \code{uniroot} function is called to solve for
the hazards, and \code{int} is the initial bracketing interval used in the
search.  Since the magnitude of the hazard functions are dependent on the
time units, it may sometimes be necessary to modify this interval.

Given a specified ratio, \code{correl.power} computes the power.  Given a
specified power, \code{correl.ratio} computes the corresponding hazard ratio
(two ratios are computed, one for the difference in each direction).
}
\examples{
correl.power(355,172,1.54,.4,7,4)
correl.power(355,172,3.36,.92,7,4)
correl.power(355,172,1.67,.7,7,4)

# check symmetry
correl.power(355,172,2/3,.5,7,4)
correl.power(355,172,1.5,.5,7,4)
correl.power(355,172,2/3,.25,7,4)
correl.power(355,172,1.5,.75,7,4)

correl.ratio(1000,150,.2,2,3.635,int=c(.0001,20))

}
\references{
Rubinstein, Gail, Santner (1981). \emph{J Chron Dis} \strong{8}: 67-74.
Rubinstein, Gail, Santner (1981). \emph{J Chron Dis} \strong{34}: 469-479.
}
\seealso{
\code{\link{powlgrnk}}
}
\keyword{design}
\keyword{survival}
