% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desmon-package.R
\docType{package}
\name{dsmn}
\alias{dsmn}
\title{Functions for design and monitoring of clincal trials}
\description{
Performs various calculations related to design of studies with either
failure time or binary endpoints, and also functions to facilitate
monitoring and analysis of such studies, including group sequential
studies with failure time endpoints and two stage and randomized phase
II studies.
}
\author{
Bob Gray \email{gray@jimmy.harvard.edu}
}
