% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nevents.R
\name{nevents.cure}
\alias{nevents.cure}
\title{Calculates the expected number of events for cure rate failure models}
\usage{
nevents.cure(
  acc.rate,
  acc.per,
  add.fu,
  cure.rate,
  haz = NULL,
  med = NULL,
  haz.ratio = 1
)
}
\arguments{
\item{acc.rate}{The accrual rate in patients per time unit}

\item{acc.per}{Duration of accrual}

\item{add.fu}{Duration of follow-up from end of accrual to analysis}

\item{cure.rate}{Proportion cured}

\item{haz}{Exponential hazard rate for failure times for those not cured.
Not needed if \code{med} is specified.}

\item{med}{Median failure time for those not cured.  Not needed if
\code{haz} is specified.}

\item{haz.ratio}{Proportional hazards shift from the exponential cure rate
model}
}
\value{
A vector of length 3 giving the expected number of failures
\code{nevents}, the failure probability \code{fail.prob} and the total
sample size \code{n}.
}
\description{
Given a uniform accrual period, specified accrual rate and follow-up period,
and specified parameters for the exponential cure rate failure model,
computes the expected number of failures
}
\details{
The exponential cure rate model has survivor function S(t) equal to
\code{p+(1-p)*exp(-l*t)}, where \code{p = cure.rate} and \code{l = haz}.
\code{haz} and \code{med} are related through \code{haz=log(2)/med}.
Optionally, a proportional hazards shift from the cure rate model with
survivor function given by \code{(p+(1-p)*exp(-l*t))^a}, where \code{a =
haz.ratio}, can be used by specifying a value for \code{haz.ratio}.  The
calculations assume that \code{n=acc.rate*acc.per} patients are entered
uniformly over the period \code{[0,acc.per]}, with follow-up continuing for
an addition \code{add.fu} time units, so censoring will be uniform on
\code{[add.fu,add.fu+acc.per]}.  The failure probability under the specified
failure distribution is then computed, as is the expected number of
failures, \code{n*fail.prob}.

Caution: consistent time units must be used for all quantities.
}
\examples{
acc.per <- 2.2
add.fu <- 3.2
nevents.cure(240,acc.per,add.fu,cure.rate=.38,med=1.7)
nevents.cure(240,acc.per,add.fu,cure.rate=.38,med=1.7,haz.ratio=.667)

}
\seealso{
\code{\link{nevents}}
}
\keyword{design}
\keyword{survival}
