% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nevents.R
\name{nevents}
\alias{nevents}
\title{Given a uniform accrual period, specified accrual rate and follow-up period,
and specified parameters for the exponential failure distribution, computes
the expected number of failures}
\usage{
nevents(acc.rate, acc.per, add.fu, haz = NULL, med = NULL)
}
\arguments{
\item{acc.rate}{The accrual rate in patients per time unit}

\item{acc.per}{Duration of accrual}

\item{add.fu}{Duration of follow-up from end of accrual to analysis}

\item{haz}{Exponential hazard rate for failure times.  Not needed if
\code{med} is specified.}

\item{med}{Median failure time.  Not needed if \code{haz} is specified.}
}
\value{
A vector of length 3 giving the expected number of failures
\code{nevents}, the failure probability \code{fail.prob} and the total
sample size \code{n}.
}
\description{
Given a uniform accrual period, specified accrual rate and follow-up period,
and specified parameters for the exponential failure distribution, computes
the expected number of failures
}
\details{
The exponential distribution has survivor function S(t) equal to
\code{exp(-l*t)}, where \code{l = haz}.  \code{haz} and \code{med} are
related through \code{haz=log(2)/med}.  The calculations assume that
\code{n=acc.rate*acc.per} patients are entered uniformly over the period
\code{[0,acc.per]}, with follow-up continuing for an addition \code{add.fu}
time units, so censoring will be uniform on \code{[add.fu,add.fu+acc.per]}.
The failure probability under the specified failure distribution is then
computed, as is the expected number of failures, \code{n*fail.prob}.

Caution: consistent time units must be used for all quantities.
}
\examples{
acc.per <- 2.2
add.fu <- 3.2
nevents(240,acc.per,add.fu,med=1.7)

}
\seealso{
\code{\link{nevents.cure}}
}
\keyword{design}
\keyword{survival}
