% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2.R
\name{onearm}
\alias{onearm}
\title{One Sample Binomial Design}
\usage{
onearm(n, p0, pa, alpha = 0.1)
}
\arguments{
\item{n}{sample size}

\item{p0}{The null response probability}

\item{pa}{The alternative response probability}

\item{alpha}{The one-sided type I error rate}
}
\value{
A vector giving the critical value and the rejection probabilities
under the null and alternative.
}
\description{
Determines the critical value and power for an exact one sample binomial
test
}
\details{
Let \code{R} be the number of responses.  If \code{p0<pa}, determines the
smallest critical value \code{r} such that \code{P(R>=r|p0)<=alpha}, and
computes \code{P(R>=r|p)} for \code{p=p0,pa}.  If \code{p0>pa}, similar
calculations are done for rejecting in the opposite tail.
}
\examples{
onearm(30, 0.1, 0.3)

}
\seealso{
\code{\link{bin1samp}}
}
\keyword{design}
