% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2.R
\name{pickwin}
\alias{pickwin}
\title{Two-Stage Randomized Phase II Designs}
\usage{
pickwin(n1, n2, p1, p2, r1, r2, r3 = r1, r4 = r2)
}
\arguments{
\item{n1}{first stage sample size (assumed to be the same for both
treatments)}

\item{n2}{addition accrual during the second stage (assumed to be the same
for both treatments)}

\item{p1}{True response probability on treatment 1}

\item{p2}{True response probability on treatment 2}

\item{r1}{The maximum number of responses to declare treatment 1 inactive in
the first stage}

\item{r2}{The maximum number of total responses over both stages to declare
treatment 1 inactive}

\item{r3}{The maximum number of responses to declare treatment 2 inactive in
the first stage}

\item{r4}{The maximum number of total responses over both stages to declare
treatment 2 inactive}
}
\value{
Returns a vector of length 9, whose components are (in order)
P(X1<=r1, X2>r4), P(r1<X1<=r2, X2>r4), P(X1>r2, X2>X1, X2>r4), P(X2<=r3,
X1>r2), P(r3<X2<=r4, X1>r2), P(X1>r2, X2>r4, X1>X2), P(X1<=r2, X2<=r4),
P(X1>r2, X2>r4, X1=X2), and the sum of the first 8 terms.
}
\description{
Calculates operating characteristics of pick the winner rules for two stage
randomized phase II designs
}
\details{
Assumes that during the first stage of accrual \code{n1} patients are
randomized to each treatment (RX1 and RX2).  If more than \code{r1}
(\code{r3}) responses are seen on treatment RX1 (RX2), then an additional
\code{n2} patients will be accrued to treatment RX1 (RX2).  If more than
\code{r2} (\code{r4}) total responses are seen on RX1 (RX2), then RX1 (RX2)
will be declared active.

Let X1 and X2 be the number of responses observed on treatments 1 and 2.
This function computes and prints P(X1<=r2)=P(RX1 declared inactive),
P(X2<=r4)=P(RX2 declared inactive), P(X2>r4 and X2>X1)=P(RX2 declared
superior), P(X1>r2 and X1>X2)=P(RX1 declared superior), and P(X1=X2 or
(X1<=r2 and X2<=r4))=P(neither declared superior).  Note that P(neither
declared superior) includes both the probability that both are declared
inactive and the probability that both are declared active and are equal.
}
\examples{
pickwin(14, 18, 0.1, 0.2, 0, 3)

}
\references{
Simon R, Wittes RE and Ellenberg SS, 1985. Randomized phase II
clinical trials. \emph{Cancer Treat Rep} \strong{69} (12):1375--81.
}
\seealso{
\code{\link{b2p}}
}
\keyword{design}
