% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin1samp.R
\name{print.bin1samp}
\alias{print.bin1samp}
\title{Print a summary of the output from bin1samp}
\usage{
\method{print}{bin1samp}(x, ...)
}
\arguments{
\item{x}{An object (vector) of class \code{bin1samp}}

\item{...}{Included for compatibility with the generic function}
}
\value{
No value is returned - used for side-effect of printing to console.
}
\description{
Prints a summary of the output from \code{\link{bin1samp}}.
}
\details{
Prints operating characteristics and stopping rules of
\code{\link{bin1samp}}.
}
\seealso{
\code{\link{bin1samp}}
}
