% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powlgrnk6.R
\name{print.powlgrnk6}
\alias{print.powlgrnk6}
\title{Print a summary of the output from powlgrnk6}
\usage{
\method{print}{powlgrnk6}(x, ...)
}
\arguments{
\item{x}{A matrix of class \code{\link{powlgrnk6}}}

\item{...}{Not used}
}
\value{
No value is returned - used for side-effect of printing to console.
}
\description{
Prints a summary of the output from \code{\link{powlgrnk6}}.
}
\details{
Prints the proportion of samples rejecting the null, the proportion stopping
in favor of the null, the average information and calendar times at
termination, and the distribution of the total number of analyses.
}
\seealso{
\code{\link{powlgrnk6}}
}
\keyword{design}
\keyword{survival}
