% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostg.R
\name{print.twostg}
\alias{print.twostg}
\title{Print a summary of the output from twostg}
\usage{
\method{print}{twostg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{twostg} (output of \code{twostg})}

\item{...}{included for compatibility with generic.  Any additional
arguments are ignored.}
}
\value{
No value is returned - used for side-effect of printing to console.
}
\description{
Prints a summary of the output from \code{\link{twostg}}
}
\details{
Prints operating characteristics and stopping rules of \code{\link{twostg}}
}
\seealso{
\code{\link{twostg}}
}
