% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2.R
\name{rp21}
\alias{rp21}
\title{Outcome Probabilities for Randomized Phase II Designs}
\usage{
rp21(n, p1, p2, crit = 0)
}
\arguments{
\item{n}{Planned number of subjects per arm}

\item{p1}{Success probability for treatment 1}

\item{p2}{Success probability for treatment 2}

\item{crit}{Maximum difference in number of responses that would be viewed
as an equivalent outcome}
}
\value{
A vector giving P(R1>R2+crit), P(|R1-R2|<=crit) and P(R1<R2-crit).
}
\description{
For a single stage randomized phase II study, computes the probability that
the difference in the number of responses is larger than a specified
critical value.
}
\details{
Computations assume two independent binomials R1 and R2, with R1 distributed
Binomal(n,p1) and R2 distributed Binomial(n,p2)
}
\examples{
rp21(32, 0.2, 0.1)

}
\seealso{
\code{\link{pickwin}}
}
\keyword{design}
