% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostg.R
\name{simon}
\alias{simon}
\title{Optimal Two-Stage Single-Arm Designs}
\usage{
simon(
  p0,
  pa,
  n1max = 0,
  ntmax = 1e+05,
  alpha = 0.1,
  beta = 0.1,
  del = 1,
  minimax = FALSE
)
}
\arguments{
\item{p0}{Null hypothesis response probability}

\item{pa}{Alternative hypothesis response probability}

\item{n1max}{The maximum number of subjects entered during the first stage.
Ignored if \code{<= 0}.}

\item{ntmax}{The maximum total number of subjects.}

\item{alpha}{Type I error rate}

\item{beta}{Type II error rate}

\item{del}{Searches for designs where the expected number of subjects under
the null is within \code{del} of the minimum possible value}

\item{minimax}{If \code{TRUE}, only searches for designs with the total
sample size equal to the minimum possible value}
}
\value{
Returns a list with components \item{designs }{A matrix with a row
giving a summary of each design meeting the criteria.  The columns are
\code{n1}, the number of subjects entered in the first stage; \code{r1}, the
cutoff for stopping at the first stage (continue if \# responses \code{>
r1}); \code{n2}, the additional number of subjects enrolled in the second
stage; \code{r2}, the cutoff for inactivity after the second stage (reject
null if \# responses \code{> r2}); \code{Pstop1.H0}, the probability of
stopping after the first stage under H0; \code{size}, the actual type I
error; \code{type2}, the actual type II error; \code{E.tot.n.H0}, the
expected \# subjects under H0.} \item{call }{The call to \code{simon()}.}
\item{description }{A text string giving a brief description of the columns
in \code{$designs}.}
}
\description{
For studies with binary endpoints, searches for two-stage sampling designs
that minimize the expected number of subjects under the null, subject to
various constraints.
}
\details{
For two-stage phase II designs for studies with binary endpoints, searches
over possible two-stage sampling designs to find those that minimize the
expected \# of subjects, subject to specified constraints. If the only
constraints are the type I and type II errors of the tests, then the designs
are the optimal designs of Simon (1989).  If a positive value of
\code{n1max} is specified, then only designs with \code{<= n1max} subjects
in the first stage are considered.  Also, only designs with \code{<= ntmax}
total subjects are considered.  Setting \code{ntmax} to a large value (as in
the default), effectively allows the search to consider all possible
designs.

If \code{minimax=TRUE}, then the minimax designs of Simon (1989), which
minimize the maximum sample size, are considered.  As there are typically
multiple designs with the same minimum max sample size, the program still
selects among the designs in this class based on the the expected sample
size under the null.

Designs which optimize one particular criterion sometimes have other
undesirable properties.  By specifying a value of \code{del > 0}, alternate
designs meeting the criteria that have expected sample sizes under the null
within \code{del} of the optimal design will also be retained.
}
\examples{
simon(0.05, 0.20)
simon(0.05, 0.20, minimax = TRUE)

}
\references{
Simon R (1989). Optimal two-stage designs for phase II clinical
trials. \emph{Controlled Clinical Trials} \strong{10}:1-10.
}
\seealso{
\code{\link{twostg}}; \code{\link{bin1samp}}; \code{\link{pickwin}};
\code{\link{rp21}}
}
\keyword{design}
