% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv1sel.R
\name{surv1sel}
\alias{surv1sel}
\title{Compute selection probabilities in a randomized phase II study with a
failure time endpoint}
\usage{
surv1sel(ng, npg, acc.per, nevents, haz, nsamp = 1000)
}
\arguments{
\item{ng}{Number of groups}

\item{npg}{Number of subjects per group (the same scalar value is assumed
for all groups)}

\item{acc.per}{The planned accrual period}

\item{nevents}{The total planned number of events (across all arms) at the
time of analysis}

\item{haz}{A vector of length \code{ng} giving the failure hazard rates for
the expoential distribution in each group}

\item{nsamp}{The number of samples to generate}
}
\value{
The vector giving the proportion of times each group is selected as
the best (ie has the lowest estimated hazard rate).
}
\description{
Uses simulations with exponential failure times and uniform accrual to
estimate the probability of each arm being the best in a randomized phase II
study, where best is defined as the lowest hazard rate in a Cox proportional
hazards model.
}
\details{
Assumes that the clinical trial will enroll \code{ng*npg} subjects uniformly
over the period \code{(0,acc.per)}, with the analysis performed when
\code{nevents} failures have been observed.  A Cox proportional hazards
model with indicator variables for the groups is fit and the group estimated
to have the lowest hazard rate is selected as the best.  In the function,
samples are repeatedly generated from exponential distributions with the
specified hazard rates and the proportion of the samples for which each
group is selected as the best is calculated.
}
\examples{
surv1sel(6, 55, 12, 40 * 6, log(2) / c(7.2, 4.8, 4.8, 4.8, 4.8, 4.8))

}
\seealso{
\code{\link{surv1samp}}
}
\keyword{design}
\keyword{survival}
