% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2.R
\name{twocon}
\alias{twocon}
\title{Confidence Interval on the Response Rate from a Two Stage Study}
\usage{
twocon(n1, n2, r1, r, conf = 0.95, dp = 1)
}
\arguments{
\item{n1}{Number of cases entered during the first stage}

\item{n2}{Number of additional cases to be entered during the second stage}

\item{r1}{max number of responses that can be observed in the first stage
without continuing}

\item{r}{total number responses observed}

\item{conf}{two-sided confidence level (proportion) for the confidence
interval}

\item{dp}{Affects the ordering of outcomes within the sample space (see
below)}
}
\value{
A vector with the lower confidence limit, the upper confidence
limit, the bias corrected MLE, the MLE, and the unbiased estimator.
}
\description{
Computes a confidence interval and several estimators of the response rate
using data from a phase II study with two-stage sampling
}
\details{
First \code{n1} patients are entered on the study.  If more than \code{r1}
responses are observed, then an additional \code{n2} patients are entered.
This function assumes that if the observed number of response \code{r < r1},
then only \code{n1} patients were entered.

The estimators computed are the MLE (the observed proportion of responses),
a bias corrected MLE, and an unbiased estimator, which is sometimes
incorrectly described as the UMVUE.

The confidence interval is based on the exact sampling distribution.
However, there is not a universally accepted ordering on the sample space in
two-stage designs.  The parameter \code{dp} can be used to modify the
ordering by weighting points within the sample space differently.
\code{dp=0} will give the Atkinson and Brown procedure, and \code{dp=1} will
order outcomes base on the MLE.  The Atkinson and Brown procedure orders
outcomes based solely on the number of responses, regardless of the number
cases sampled.  The MLE ordering defines as more extreme those outcomes with
a more extreme value of the MLE (the proportion of responses).  Other powers
of \code{dp}, such as \code{dp=1/2}, could also be used.  Let \code{R} be
the number of responses and \code{N=n1} if \code{R<=r1} and \code{N=n1+n2}
if \code{R>r1}.  In general, the outcomes that are more extreme in the high
response direction are those with \code{R/(N^dp) >= r/(n^dp)}, where
\code{r} and \code{n} are the observed values of \code{R} and \code{N}, and
the outcomes that are more extreme in the low response direction are those
with \code{R/(N^dp) <= r/(n^dp)}.
}
\examples{
twocon(14, 18, 3, 4, dp = 0)
twocon(14, 18, 3, 4, dp = 1)

}
\references{
Atkinson and Brown (1985), BIOMETRICS 741-744.
}
\seealso{
\code{\link{binci}}
}
\keyword{htest}
