% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostg.R
\name{twostg}
\alias{twostg}
\title{Operating Characteristics of Two-Stage Phase II Designs}
\usage{
twostg(n1, n2, p1, r1, r2)
}
\arguments{
\item{n1}{Number of cases accrued in the first stage}

\item{n2}{Number of additional cases accrued in the second stage}

\item{p1}{Response probability}

\item{r1}{max number responses in first stage where drug would still be
declared ineffective}

\item{r2}{max number of total responses for drug to be declared ineffective}
}
\value{
\code{twostg} returns a list of class \code{twostg}, with components
\item{inputs }{a vector containing the input values} \item{prob.inactive}{a
vector giving the total probability of \code{r2} or fewer responses and the
probability of \code{r1} or fewer in the first stage}

\code{print.twostg} prints a summary of the input data and the probability
of being declared ineffective overall and at the first stage.
}
\description{
Calculates the operating characteristics of single arm, two-stage phase II
designs.
}
\details{
In the first stage, \code{n1} cases are accrued and the study is stopped if
\code{r1} or fewer responses are observed.  If more than \code{r1} responses
are observed then an additional \code{n2} cases are accrued.  The drug or
regimen will be declared ineffective if \code{r2} or fewer total responses
are observed, and worth further study if more than \code{r2} responses are
observed.  Subjects are assumed to be independent with a common response
probability \code{p1}.
}
\examples{
twostg(14, 18, 0.1, 1, 4)
twostg(14, 18, 0.3, 1, 4)

}
\seealso{
\code{\link{pickwin}}, \code{\link{rp21}}, \code{\link{simon}},
\code{\link{bin1samp}}
}
\keyword{design}
