% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostg3.R
\name{twostg3}
\alias{twostg3}
\title{Compute operating characteristics of two-stage trinomial-outcome designs}
\usage{
twostg3(n1, n2, presp, pstab, r1, s1, r2, s2)
}
\arguments{
\item{n1}{Number of subjects enrolled in the first stage}

\item{n2}{Number of additional subjects enrolled in the second stage}

\item{presp}{The probability of the best (response) category}

\item{pstab}{The probability of the intermediate (stable only) category}

\item{r1}{Max number of responses that can occur at the first stage without
proceeding to the second stage}

\item{s1}{Max number cases that can have responses or stable outcomes at
first stage without proceeding to the second stage}

\item{r2}{Max number of responses from stages 1 and 2 combined that can be
observed without declaring the treatment to be effective}

\item{s2}{Max number of cases from stages 1 and 2 combined that can have
responses or stable outcomes without declaring the treatment to be effective}
}
\value{
A vector giving the probability of stopping after the first stage
(\code{p.stop.1}) and the overall probability that the treatment is declared
inactive (\code{p.inactive}).
}
\description{
For a phase II study with three possible ordered outcomes (e.g., response,
stable, progression), computes the probability of not rejecting the null
hypothesis for a specified two-stage design.
}
\details{
Consider a phase II study with an ordered categorical outcome with three
categories, which are referred to here as response, stable and progression
(based on the best response outcome in oncology).  A treatment might improve
outcome by either improving the response rate or by improving the disease
stabilization rate (where stabilization includes best response of response
or stable).  In the designs considered here, the treatment is declared to be
active (or sufficiently active to be worth studying further) if it shows a
sufficient level of activity for either response or disease stabilization.

More formally, let \code{R1} be the number of patients observed to respond
and \code{S1} be the number of patients observed to be stable (but not
responses) during the first stage of accrual, and similarly let \code{R2}
and \code{S2} be the number of additional cases observed to respond or
stabilize from the second stage.  The study is stopped after the first stage
if BOTH \code{R1 <= r1} and \code{R1+S1 <= s1}.  The study proceeds to the
second stage if either \code{R1 > r1} or \code{R1+S1 > s1}.  The treatment
is declared to be inactive if either the study stops after the first stage
or if at the end of the second stage BOTH \code{R1+R2 <= r2} and
\code{R1+R2+S1+S2 <= s2}.  The treatment is considered sufficiently active
for further investigation if either \code{R1+R2 > r2} or \code{R1+R2+S1+S2 >
s2}.

Note that \code{presp} and \code{pstab} are the probabilities of the
corresponding disjoint cells in the trinomial model (that is, \code{pstab}
is the probability that a case will be stable but not a response).
}
\examples{
twostg3(20, 40, 0.05, 0.10, 1, 3, 5, 22)

}
\seealso{
\code{\link{twostg}}
}
\keyword{design}
