% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox.pow.R
\name{wilcox.pow}
\alias{wilcox.pow}
\title{Compute Power for the Wilcoxon and T Tests}
\usage{
wilcox.pow(n1, n2 = NULL, del, alpha2 = 0.05, tdist = FALSE)
}
\arguments{
\item{n1}{sample size in group 1}

\item{n2}{Sample size in group 2. If NULL, assumes a 1-sample test.}

\item{del}{difference as a fraction of the population standard deviation}

\item{alpha2}{two-sided significance level}

\item{tdist}{if TRUE, uses t distribution for the critical value; if FALSE,
uses a normal critical value}
}
\value{
A vector of length two giving the power for the Wilcoxon (first component)
and t (second component) tests.
}
\description{
Computes an approximation to the power of the Wilcoxon signed rank or
two-sample Wilcoxon rank sum test and one- or two-sample t-tests
}
\details{
Uses an approximations given in Lehman, Nonparametrics: Statistical Methods
Based on Ranks, for the power of the Wilcoxon signed rank and two-sample
Wilcoxon test.  Assumes the data on a transformed scale is normally
distributed with constant variances.  The difference \code{del} is a
location shift on this transformed scale, expressed as a fraction of the
common standard deviation.

The critical value from the t-distribution with n1-1 (one-sample) or n1+n2-2
(two-sample) degrees of freedom can be used instead of the standard normal
by specifying \code{tdist=TRUE}.  However, the rejection probability is
still computed using the normal distribution.
}
\examples{
wilcox.pow(24, 24, 1)
wilcox.pow(24, 10, 1)
wilcox.pow(10, NULL, 0.5) ## one-sample test

}
\keyword{design}
